api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0
snowflake-connector-python[pandas]==3.12.0
cachetools==5.5.0
python-dotenv==1.0.1
httpx==0.27.0
